+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Note: Breaking changes between versions are indicated by "💥".
 
 ## Unreleased
 
+- [Feature] Add ``images printtag`` command
 - [Improvement] Make it possible to override individual contact email addresses with plugins
 - [Bugfix] Replace "no-reply@LMS_BASE" email address by regular contact email address
 - [Bugfix] Disable learner records globally by default